bind PositionsSnapshot service
